COMT modeling ECCOFS, SECOFS, NECOFS (#117)
Note from @mykelalvis - squashed since rebasing was not an option.

* minor comment fix/typo

* run_storms.sh NWS 14 section

* restrictive profile, automating useradd work

* user management

* clean up

* memorable name moved to end of tag

* secofs runs, working on fsx

* secofs schism, fsx workflows, nosofs

* eccofs tested

* eccofs tests

* necofs from lixia builds

* necofs cold run

* eccofs build and testcase input data scripts

* necofs coldstart runs

---------

+++ cloudflow/workflows/tasks.py
@@ -35,7 +35,7 @@ from cloudflow.services.S3Storage import S3Storage
 
 from cloudflow.services.ScratchDisk import ScratchDisk
 from cloudflow.services.FSxScratchDisk import FSxScratchDisk
-from cloudflow.services.NFSScratchDisk import NFSScratchDisk
+# from cloudflow.services.NFSScratchDisk import NFSScratchDisk
 
 from cloudflow.utils import modelUtil as util
 
@@ -51,7 +51,7 @@ log = logging.getLogger('workflow')
 ##############
 
 @task
-def create_scratch(provider: str, jobconfigfile: str, mountpath: str = '/ptmp') -> ScratchDisk:
+def create_scratch(provider: str, cluster: Cluster, job: Job) -> ScratchDisk:
     """ Provides a high speed scratch disk if available. Creates and mounts the disk.
 
     Parameters
@@ -70,24 +70,28 @@ def create_scratch(provider: str, jobconfigfile: str, mountpath: str = '/ptmp') 
 
     """
 
-    ### PT 2/13/2025 - HERE
+    ### calls the object __init__
     if provider == 'FSx':
-        scratch = FSxScratchDisk(configfile)
+        scratch = FSxScratchDisk(cluster, job)
     elif provider == 'NFS':
-        scratch = NFSScratchDisk(configfile)
+        scratch = NFSScratchDisk(cluster, job)
     elif provider == 'Local':
         log.error('Coming soon ...')
-        raise signals.FAIL('FAILED')
+        return
+    elif provider == 'None':
+        log.info('No scratch disk being used')
+        return
     else:
         log.error('Unsupported provider')
-        raise signals.FAIL('FAILED')
+        return
 
-    scratch.create(mountpath)
+    log.debug(f"Creating scratch drive {provider}") 
+    scratch.create()
     return scratch
 
 
 @task
-def mount_scratch(scratch: ScratchDisk, cluster: Cluster):
+def scratch_remote_mount(scratch: ScratchDisk, cluster: Cluster, job: Job):
     """ Mounts the scratch disk on each node of the cluster
 
     Parameters
@@ -98,8 +102,13 @@ def mount_scratch(scratch: ScratchDisk, cluster: Cluster):
       The cluster object that contains the hostnames and tags
     """ 
 
-    hosts = cluster.getHosts() 
-    scratch.remote_mount(hosts)
+
+    if isinstance(scratch, ScratchDisk):
+        log.info("Attempting to mount scratch disk on compute nodes...")
+        hosts = cluster.getHosts()
+        log.debug(f"cluster.hosts: {hosts}")
+        scratch.remote_mount(hosts)
+
     return
 
 
@@ -112,7 +121,9 @@ def delete_scratch(scratch: ScratchDisk):
     scratch : ScratchDisk
       The scratch disk object
     """
-    scratch.delete()
+    if isinstance(scratch, ScratchDisk):
+        scratch.delete()
+
     return
 
 # Storage
@@ -299,16 +310,77 @@ def job_init(cluster: Cluster, configfile) -> Job:
     """
     NPROCS = cluster.nodeCount * cluster.PPN
 
-    if debug: print(f"DEBUG: in tasks job_init configfile: {configfile}")
-    # This is opposite of the way resources are usually allocated
-    # Normally, NPROCs comes from the job and will use that much compute resources
-    # Here, we fit the job to the desired on-demand resources.
-    # TODO ?: set it up so that the best machine(s) for the job are provisioned based on the resource request.
+    log.debug(f"DEBUG: in tasks job_init configfile: {configfile}")
 
     factory = JobFactory()
     job = factory.job(configfile, NPROCS)
     return job
 
+# cluster, job
+@task
+def simple_run(cluster: Cluster, job: Job):
+    """ Run the forecast
+
+    Parameters
+    ----------
+    cluster : Cluster
+        The cluster to run on
+    job : Job
+        The job to run
+    """
+
+    # Easier to read
+    OFS = job.OFS
+    PPN = cluster.PPN
+    NPROCS = job.NPROCS
+
+    #CDATE = job.CDATE
+    #HH = job.HH
+    #OUTDIR = job.OUTDIR
+
+    WRKDIR = job.WRKDIR
+    RUNDIR = job.RUNDIR
+    INPUTFILE = job.INPUTFILE
+    EXEC = job.EXEC
+
+    PTMP = getattr(job, "PTMP", 'none')
+
+    try:
+        HOSTS = cluster.getHostsCSV()
+    except Exception as e:
+        log.exception('In driver: execption retrieving list of hostnames:' + str(e))
+        raise signals.FAIL('FAILED')
+
+    runscript = f"{curdir}/simple_launcher.sh"
+
+    try:
+        if OFS in ('necofs_cold', 'necofs'):
+
+        # export OFS=$1
+        # export HOSTS=$2
+        # export NPROCS=$3
+        # export PPN=$4
+        # export WRKDIR=$5
+        # export RUNDIR=$6
+        # export INPUTFILE=$7
+        # export EXEC=$8
+
+            args = [ runscript, OFS, HOSTS, str(NPROCS), str(PPN), WRKDIR, RUNDIR, INPUTFILE, EXEC ]
+            result = subprocess.run(args, stderr=subprocess.STDOUT, universal_newlines=True)
+        else:
+            raise signals.FAIL(f"ERROR: don't know how to run {OFS}")
+
+        if result.returncode != 0:
+            log.exception(f'Forecast failed ... result: {result.returncode}')
+            raise signals.FAIL(f'Forecast failed ... result: {result.returncode}')
+
+    except Exception as e:
+        log.exception('In driver: Exception during subprocess.run :' + str(e))
+        raise signals.FAIL('In driver: Exception during subprocess.run :' + str(e))
+
+    log.info('Forecast finished successfully')
+
+    return
 
 
 
@@ -324,15 +396,20 @@ def forecast_run(cluster: Cluster, job: Job):
     job : Job
         The job to run
     """
-    PPN = cluster.getCoresPN()
 
     # Easier to read
+    PPN = cluster.PPN
     CDATE = job.CDATE
     HH = job.HH
     OFS = job.OFS
     NPROCS = job.NPROCS
     OUTDIR = job.OUTDIR
     SAVEDIR = job.SAVE
+
+    PTMP = getattr(job, "PTMP", 'none')
+
+    # for secofs
+    NSCRIBES = getattr(job, "NSCRIBES", '')
 
     runscript = f"{curdir}/fcst_launcher.sh"
 
@@ -346,9 +423,9 @@ def forecast_run(cluster: Cluster, job: Job):
 
         if OFS == "adnoc":
             time.sleep(60)
-            result = subprocess.run([runscript, CDATE, HH, OUTDIR, SAVEDIR, str(NPROCS), str(PPN), HOSTS, OFS, job.EXEC], stderr=subprocess.STDOUT, universal_newlines=True)
+            result = subprocess.run([runscript, CDATE, HH, OUTDIR, SAVEDIR, PTMP, str(NPROCS), str(PPN), HOSTS, OFS, job.EXEC], stderr=subprocess.STDOUT, universal_newlines=True)
         else:
-            result = subprocess.run([runscript, CDATE, HH, OUTDIR, SAVEDIR, str(NPROCS), str(PPN), HOSTS, OFS, job.EXEC], stderr=subprocess.STDOUT, universal_newlines=True)
+            result = subprocess.run([runscript, CDATE, HH, OUTDIR, SAVEDIR, PTMP, str(NPROCS), str(PPN), HOSTS, OFS, job.EXEC, NSCRIBES], stderr=subprocess.STDOUT, universal_newlines=True)
 
         if result.returncode != 0:
             log.exception(f'Forecast failed ... result: {result.returncode}')
@@ -388,8 +465,15 @@ def hindcast_run_multi(cluster: Cluster, job: Job):
     HH = job.HH
     OFS = job.OFS
     NPROCS = job.NPROCS
-    OUTDIR = job.OUTDIR
+
     SAVEDIR = job.SAVE
+
+    PTMP = getattr(job, "PTMP", 'none')
+
+    # Use an environment variable for FSx signal
+
+    if OFS == "secofs":
+       JOBARGS = getattr(job, "NSCRIBES", '')
 
     runscript = f"{curdir}/fcst_launcher.sh"
     print(f"In hindcast_run_multi: runscript: {runscript}")
@@ -409,9 +493,17 @@ def hindcast_run_multi(cluster: Cluster, job: Job):
         # Create ocean in file
         job.make_oceanin()
 
+        if OFS == "eccofs":
+            JOBARGS = getattr(job, "OCEANIN", '')
+
+        OUTDIR = job.OUTDIR
+
         try:
             print('Launching model run ...')
-            result = subprocess.run([runscript, job.CDATE, HH, job.OUTDIR, SAVEDIR, str(NPROCS), str(PPN), HOSTS, OFS, job.EXEC], stderr=subprocess.STDOUT, universal_newlines=True)
+            # TODO: too many script levels?
+            # TODO: where should this be encapsulated? 
+            # Maybe do it in python instead of bash, can have named arguments or use args**
+            result = subprocess.run([runscript, job.CDATE, HH, OUTDIR, SAVEDIR, PTMP, str(NPROCS), str(PPN), HOSTS, OFS, job.EXEC, JOBARGS], stderr=subprocess.STDOUT, universal_newlines=True)
 
             if result.returncode != 0:
                 log.exception(f'Forecast failed ... result: {result.returncode}')
@@ -422,10 +514,13 @@ def hindcast_run_multi(cluster: Cluster, job: Job):
             raise signals.FAIL('FAILED')
 
         # Set the date for the next day run
+        # TODO: most hindcasts don't run for a single day at a time (exception LiveOcean)
+        # It depends on restart state availability and other factors
+        # Take on a case by case basis, perhaps add the increment to task argument or embed in the job
         job.CDATE = util.ndate(job.CDATE, 1)
 
     # end of while loop
-
+    return
 
 ###############################################################################
 # cluster, job
